Terminado ejercicio de WPF

+++ 16-WPFPersonas/16-WPFPersonas-UI/ViewModels/clsMainPageVM.cs
@@ -16,15 +16,20 @@ namespace _16_WPFPersonas_UI.ViewModels
         #region "Atributos"
         private clsPersona _personaSeleccionada;
         public List<clsPersona> lista { get; set; }
-        
+
         private DelegateCommand _eliminarCommand;
         private DelegateCommand _guardarCommand;
-        private DelegateCommand _insertarPersona;
+        private DelegateCommand _insertarCommand;
+        private DelegateCommand _pegarCommand;
+        private DelegateCommand _actualizarCommand;
+
+        private bool nuevo;
         #endregion
 
         public clsMainPageVM()
         {
-            lista = new clsListadosBL().getListadoPersonaBL();            
+            lista = new clsListadosBL().getListadoPersonaBL();
+            nuevo = false;
         }
 
         public clsPersona Persona
@@ -44,8 +49,9 @@ namespace _16_WPFPersonas_UI.ViewModels
 
 
 
-      
 
+
+        //Eliminar
         public DelegateCommand eliminarCommand
         {
             get
@@ -57,11 +63,16 @@ namespace _16_WPFPersonas_UI.ViewModels
 
         private void EliminarCommand_Executed()
         {
-            clsManejadoraPersonaBL mpbl = new clsManejadoraPersonaBL();
-            mpbl.deletePersona(_personaSeleccionada.id);
+            MessageBoxResult mb = MessageBox.Show("¿Seguro que quiere eliminar a " + Persona.nombre + "?", "Eliminar", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (mb == MessageBoxResult.Yes)
+            {
+                clsManejadoraPersonaBL mpbl = new clsManejadoraPersonaBL();
+                mpbl.deletePersona(_personaSeleccionada.id);
 
-            lista = new clsListadosBL().getListadoPersonaBL();
-            NotifyPropertyChanged("lista");
+                lista = new clsListadosBL().getListadoPersonaBL();
+                NotifyPropertyChanged("lista");
+            }
+
         }
 
         private bool EliminarCommand_CanExecute()
@@ -72,13 +83,29 @@ namespace _16_WPFPersonas_UI.ViewModels
             return sePuedeBorrar;
         }
 
+        //Actualizar
+        public DelegateCommand actualizarCommand
+        {
+            get
+            {
+                _actualizarCommand = new DelegateCommand(ActualizarCommand_Executed);
+                return _actualizarCommand;
+            }
+        }
 
+        private void ActualizarCommand_Executed()
+        {
+            lista = new clsListadosBL().getListadoPersonaBL();
+            nuevo = false;
+            NotifyPropertyChanged("lista");
+        }
 
+        //Guardar
         public DelegateCommand guardarCommand
         {
             get
-            {                
-               _guardarCommand= new DelegateCommand(GuardarCommand_Executed);
+            {
+                _guardarCommand = new DelegateCommand(GuardarCommand_Executed);
                 return _guardarCommand;
             }
         }
@@ -86,10 +113,54 @@ namespace _16_WPFPersonas_UI.ViewModels
         private void GuardarCommand_Executed()
         {
             clsManejadoraPersonaBL mpbl = new clsManejadoraPersonaBL();
-            mpbl.updatePersona(_personaSeleccionada);
+            if (_personaSeleccionada.id == -1)
+                mpbl.insertPersona(_personaSeleccionada);
+            else
+                mpbl.updatePersona(_personaSeleccionada);
 
             lista = new clsListadosBL().getListadoPersonaBL();
             NotifyPropertyChanged("lista");
+            _personaSeleccionada = null;
+        }
+
+
+
+
+        //Insertar
+        public DelegateCommand insertarCommand
+        {
+            get
+            {
+                _insertarCommand = new DelegateCommand(InsertarCommand_Executed);
+                return _insertarCommand;
+            }
+        }
+
+        private void InsertarCommand_Executed()
+        {
+            clsPersona p = new clsPersona("Persona", "sin guardar", -1, DateTime.Today, "", "");
+            lista = new clsListadosBL().getListadoPersonaBL();
+            lista.Add(p);
+
+            nuevo = true;
+            _personaSeleccionada = p;
+            NotifyPropertyChanged("lista");
+            NotifyPropertyChanged("Persona");
+        }
+
+        //Pegar
+        public DelegateCommand pegarCommand
+        {
+            get
+            {
+                _pegarCommand = new DelegateCommand(PegarCommand_Executed);
+                return _pegarCommand;
+            }
+        }
+
+        private void PegarCommand_Executed()
+        {
+            Application.Current.Shutdown();
         }
 
     }
